+++ db/query/exchange_rate/list_exchange_rates.sql
@@ -1,3 +1,3 @@
 -- name: ListExchangeRates :many
-SELECT id, from_currency, to_currency, rate::float8, created_at FROM exchange_rates
+SELECT id, from_currency, to_currency, rate, created_at FROM exchange_rates
 ORDER BY from_currency, to_currency;
